% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogis.R
\name{dloglogis}
\alias{dloglogis}
\title{The Log-Logistic Distribution}
\usage{
dloglogis(x, alpha, beta)
}
\arguments{
\item{beta}{}
}
\value{
frequency of the value x in a log-logistic distribution with parameters alpha and beta
}
\description{
The Log-Logistic Distribution
}
\examples{
dloglogis(1, 3, 5)


}
