% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton-raphson.R
\name{loglogistic_hessian}
\alias{loglogistic_hessian}
\title{Log-Logistic Hessian function}
\usage{
loglogistic_hessian(x, n, alpha, beta)
}
\arguments{
\item{beta}{}
}
\description{
Log-Logistic Hessian function
}
