% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton-raphson.R
\name{NR_fit_LogLogistic_hessian}
\alias{NR_fit_LogLogistic_hessian}
\title{Perform a Newton Raphson fit on the Log-Logistic distribution}
\usage{
NR_fit_LogLogistic_hessian(x, params_0, eps = 1e-05)
}
\arguments{
\item{eps}{}
}
\description{
Perform a Newton Raphson fit on the Log-Logistic distribution
}
