% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogis.R
\name{ploglogis}
\alias{ploglogis}
\title{The Log-Logistic Distribution}
\usage{
ploglogis(q, alpha, beta)
}
\arguments{
\item{beta}{}
}
\value{
probability of the quantile q in a log-logistic distribution with parameters alpha and beta
}
\description{
The Log-Logistic Distribution
}
\examples{
ploglogis(1, 1, 3)
}
