% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogis.R
\name{qloglogis}
\alias{qloglogis}
\title{The Log-Logistic Distribution}
\usage{
qloglogis(p, alpha, beta)
}
\arguments{
\item{beta}{}
}
\value{
quantile with the probability p in a log-logistic distribution with parameters alpha and beta
}
\description{
The Log-Logistic Distribution
}
\examples{
qloglogis(0.25, 1, 3)
}
