% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogis.R
\name{rloglogis}
\alias{rloglogis}
\title{The Log-Logistic Distribution}
\usage{
rloglogis(n, alpha, beta)
}
\arguments{
\item{beta}{}
}
\value{
pseudo-randomly generated dataset following the log-logistic distribution with parameters alpha and beta
}
\description{
The Log-Logistic Distribution
}
\examples{
rloglogis(1000, 3, 5)
}
